backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
fastapi==0.115.12
h11==0.14.0
idna==3.10
pydantic==2.11.3
pydantic_core==2.33.1
python-chess==1.999
sniffio==1.3.1
starlette==0.46.1
stockfish==3.28.0
typing-inspection==0.4.0
typing_extensions==4.13.1
uvicorn==0.34.0
